Fix build errors (wave 4): 762 -> 354 errors
- PacketWriter constructor: removed invalid arguments (33 files)
- WeaponType enum: CrossBow->Crossbow, Gun->Pistol/LongGun, Staff->OneHandStaff/TwoHandedStaff
- SC_LEAVE_SIGHT_PC: use leave_zone.objInstID pattern
- LogLevel.Info -> LogLevel.Information (26 files)
- SC_CHANGED_HP/MP: hp->current, mp->current field names
- Math.Clamp: explicit casts to resolve ambiguity
- ToSizedPacket: removed args, added WriteHeader calls
- HealingService: short to ushort casts
- ModCombatJobStatus: added typeID and statusPoint properties
- EQUIPPED: added AsArray() method
- String to byte[] conversions with Encoding.ASCII.GetBytes()

+++ kakia_lime_odyssey_server/Services/Combat/CombatValidator.cs
@@ -324,13 +324,13 @@ public static class CombatValidator
 			status.MeleeAttack.Def = 0;
 
 		// Clamp crit rate
-		status.MeleeAttack.CritRate = Math.Clamp(status.MeleeAttack.CritRate, 0, MaxCritRate);
+		status.MeleeAttack.CritRate = Math.Clamp((int)status.MeleeAttack.CritRate, 0, MaxCritRate);
 
 		// Clamp hit rate
-		status.MeleeAttack.Hit = Math.Max(status.MeleeAttack.Hit, 1);
+		status.MeleeAttack.Hit = (ushort)Math.Max((int)status.MeleeAttack.Hit, 1);
 
 		// Clamp flee rate
-		status.MeleeAttack.FleeRate = Math.Max(status.MeleeAttack.FleeRate, 0);
+		status.MeleeAttack.FleeRate = (ushort)Math.Max((int)status.MeleeAttack.FleeRate, 0);
 
 		return status;
 	}
